Scat based on vel_perturb

+++ PHONON1D/SRC/VEL_MODEL_CHECKS_isoscat.f90
@@ -4,7 +4,7 @@
       
       IMPLICIT NONE
       
-      REAL(8) qdepdiff
+      REAL qdepdiff
       
       !MODEL CHECKS
 			check_scat = 1
